feat: add renderer.js
